requirements.txt
flask==3.0.0
flask-cors==4.0.0
segyio==1.9.13
numpy==1.24.3
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0 